Changes in findPointersInExpression:
For pointers that are in the form of array element references, if the array type is an array of length 0 then doest report such a exception in this function but stop seeking pointers recursively. 
If that kind of pointers are dereferenced in the program, it will be reported at execution time.